Struct libva::display::Display

pub struct Display {
    handle: *mut c_void,
    drm_file: File,
}

An owned VADisplay

Fields
handle: *mut c_void

The handle to interact with the underlying VADisplay.

drm_file: File

The DRM file that must be kept open while the VADisplay is in use.

Implementations

impl Display

pub fn open() -> Result<Self>

Opens and initializes an owned Display

fn drm_open() -> Result<Option<(*mut c_void, File)>>

pub(crate) fn handle(&self) -> *mut c_void

Returns the associated handle

pub fn query_config_profiles(&self) -> Result<Vec<Type>>

Queries supported profiles

pub fn query_vendor_string(&self) -> Result<String, &'static str>

Returns a String describing some aspects of the VA implemenation on a specific hardware accelerator. The format of the returned string is vendor specific and at the discretion of the implementer. e.g. for the Intel GMA500 implementation, an example would be: “Intel GMA500 - 2.0.0.32L.0005”

pub fn query_config_entrypoints(&self, profile: Type) -> Result<Vec<Type>>

Query supported entrypoints for a given profile

pub fn get_config_attributes(
&self,
profile: Type,
entrypoint: Type,
attributes: &mut Vec<VAConfigAttrib>
) -> Result<()>

Get attributes for a given profile/entrypoint pair

pub fn create_surfaces(
self: &Rc<Self>,
rt_format: u32,
va_fourcc: Option<u32>,
width: u32,
height: u32,
usage_hint: Option<UsageHint>,
num_surfaces: u32
) -> Result<Vec<Surface>>

Create VASurfaces by wrapping around a vaCreateSurfaces call rt_format is the desired surface format. See VA_RT_FORMAT_* va_fourcc is the desired pixel format. See VA_FOURCC_* width is the surface width height is the surface height usage_hint gives the driver a hint of intended usage to optimize allocation (e.g. tiling) num_surfaces is the number of surfaces to create

pub fn create_context(
self: &Rc<Self>,
config: &Config,
coded_width: i32,
coded_height: i32,
surfaces: Option<&Vec<Surface>>,
progressive: bool
) -> Result<Context>

Create a Context by wrapping around a vaCreateContext call. config is the configuration for the context coded_width is the coded picture width coded_height is the coded picture height progressive is whether only progressive frame pictures are present in the sequence surfaces are a hint for the amount of surfaces tied to the context

pub fn create_config(
self: &Rc<Self>,
attrs: Option<Vec<VAConfigAttrib>>,
profile: Type,
entrypoint: Type
) -> Result<Config>

Create a VAConfig by wrapping around the vaCreateConfig call. attrs describe the attributes to set for this config. A list of the supported attributes for a given profile/entrypoint pair can be retrieved using Display::get_config_attributes. Other attributes will take their default values

pub fn query_image_formats(self: &Rc<Self>) -> Result<Vec<VAImageFormat>>

A wrapper over vaQueryImageFormats.