Module libva::picture

Modules

private

Structs

Picture

An abstraction over VABuffers and a VASurface suitable for decoding with vaBeginPicture, vaRenderPicture, vaEndPicture “surface” will have valid raw picture data after “begin”, “render”, “end” and “sync” are called, in this order.

PictureInner

Enums

PictureBegin

Represents a Picture after vaBeginPicture has been called.

PictureEnd

Represents a Picture after vaEndPicture has been called.

PictureNew

Represents a Picture that has just been created.

PictureRender

Represents a Picture after vaRenderPicture has been called.

PictureSync

Represents a Picture after vaSyncSurface has been called on the underlying surface.

Traits

PictureReclaimableSurface

Represents a state where one can reclaim the underlying Surface for this Picture. This is true when either no decoding has been initiated or, alternatively, when the decoding operation has completed for the underlying vaSurface

PictureState

A Picture will only have valid YUV data after a sequence of operations are performed in a particular order. This order correspond to the following VA-API calls: vaBeginPicture, vaRenderPicture, vaEndPicture and vaSyncSurface. This trait enforces this ordering by implementing the Typestate pattern to constrain what operations are available in what particular states.